FlightTaskManualAltitude: change setpoint instead of constraint
Makes more sense from a control perspective to change the
setpoint directly whenever you have direct access to it instead
of constraining later on. The constraints were introduced to
be able to constrain e.g. the velocity when commanding a
position setpoint and not having access to the resulting
velocity setpoint.